LanguageAssistant-backend/requirements.txt
uvicorn
fastapi
pydantic

load_dotenv

llama_index.core
llama_index.llms.litellm
llama_index.embeddings.huggingface
